# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DEHQON",
        "version": "0.1"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "description": "Register a new user and return an access token upon successful registration.",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "surname",
                                    "phone",
                                    "born_in",
                                    "password",
                                    "gender",
                                    "currency",
                                    "register_token"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "User's first name",
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "surname": {
                                        "description": "User's last name",
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "phone": {
                                        "description": "User's phone number in the format 992 XXXXXXXXX",
                                        "type": "string",
                                        "example": "992123456789"
                                    },
                                    "born_in": {
                                        "description": "User's birth date in YYYY-MM-DD format",
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "gender": {
                                        "description": "User's gender (e.g., 1 for male, 2 for female)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "currency": {
                                        "description": "Preferred currency code",
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "image": {
                                        "description": "Optional user's profile image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "register_token": {
                                        "description": "Registration token to verify the user",
                                        "type": "string",
                                        "example": "abcdef123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or invalid token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid token"
                                        },
                                        "errors": {
                                            "description": "Validation errors",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/send-code": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send code to phone number",
                "description": "This endpoint sends a code to the specified phone number for authentication purposes.",
                "operationId": "sendCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "description": "Phone number in the format 992 XXXXXXXXX",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Code sent successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "description": "Validation errors",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/check-phone/{phone}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Check if a phone number exists in the database",
                "description": "Returns a message indicating whether the phone number exists in the database.",
                "operationId": "checkPhoneNumber",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "path",
                        "description": "Phone number to check",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "992XXXXXXXXX"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Phone number exists in the database",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Phone number exists in the database."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Phone number does not exist in the database",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Phone number does not exist in the database."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid phone number format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid phone number format."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/verify-code": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify the code sent to the user's phone",
                "description": "This endpoint verifies the code sent to the user's phone and issues a registration token or an access token if the user is already registered.",
                "operationId": "verifyCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "code"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "Phone number in the format 992XXXXXXXXX",
                                        "type": "string",
                                        "example": "992123456789"
                                    },
                                    "code": {
                                        "description": "Verification code sent to the phone",
                                        "type": "integer",
                                        "example": 123456
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Code verified successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Verified"
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                        },
                                        "register_token": {
                                            "type": "string",
                                            "example": "sample_register_token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid code or validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid code"
                                        },
                                        "errors": {
                                            "description": "Validation errors",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}